[ANDROSDK-444] Pmd
